# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking:
    properties:
      check_in:
        type: string
      check_out:
        type: string
      created_at:
        type: string
      guests:
        type: integer
      id:
        type: integer
      room_id:
        type: integer
      status:
        type: string
      total_price:
        type: number
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: |-
    RESTful API for a Hotel Booking System built with Go, Gin, GORM, PostgreSQL, and JWT authentication.

    Features include authentication, hotel management, room management, bookings, availability, reviews, favorites, and payments.
  title: Hotel Booking System API
  version: "1.0"
paths:
...
  /bookings:
    get:
      description: Get all bookings. Admin and manager users only.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all bookings
      tags:
      - Bookings
    post:
      consumes:
      - application/json
      description: Create a new hotel room booking for the authenticated user.
      parameters:
      - description: Booking data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a booking
      tags:
      - Bookings
  /bookings/{id}:
    delete:
      description: Cancel an existing booking belonging to the authenticated user.
        The booking is marked as cancelled instead of being deleted.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cancel a booking
      tags:
      - Bookings
    get:
      description: Get a booking by ID. Customers can only view their own bookings,
        while admins and managers can view any booking.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get booking by ID
      tags:
      - Bookings
    put:
      consumes:
      - application/json
      description: Update an existing booking. Customers can only update their own
        bookings.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated booking data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a booking
      tags:
      - Bookings
...
  /bookings/{id}/status:
    patch:
      consumes:
      - application/json
      description: Confirm or cancel a booking. Admin and manager users only.
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      - description: Booking status
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update booking status
      tags:
      - Bookings
  /bookings/my:
    get:
      description: Get all bookings belonging to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_Mosteben_hotel-booking-system_internal_booking_model.Booking'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my bookings
      tags:
      - Bookings
...
  /rooms/{id}/availability:
    get:
      description: Check whether a room is available for the specified date range.
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: integer
      - description: Check-in date in YYYY-MM-DD format
        in: query
        name: check_in
        required: true
        type: string
      - description: Check-out date in YYYY-MM-DD format
        in: query
        name: check_out
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check room availability
      tags:
      - Room Availability
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"